requirements.txt
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
black==23.3.0
boto3==1.26.162
botocore==1.29.162
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==41.0.1
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
future==0.18.3
idna==3.4
importlib-metadata==4.13.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
junos-eznc==2.6.7
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mypy-extensions==0.4.4
napalm==4.1.0
ncclient==0.6.13
netaddr==0.8.0
netmiko==4.2.0
netutils==1.4.1
nornir==3.3.0
nornir-inspect==1.0.3
nornir-napalm==0.3.0
nornir-netbox==0.3.0
nornir-netmiko==0.2.0
nornir-utils==0.2.0
ntc-templates==3.4.0
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
paramiko==3.2.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyeapi==1.0.0
Pygments==2.15.1
PyNaCl==1.5.0
pyparsing==3.1.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rich==12.6.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
scp==0.14.5
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
textfsm==1.1.2
tomli==2.0.1
traitlets==5.9.0
transitions==0.9.0
ttp==0.9.4
ttp-templates==0.3.5
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
XlsxWriter==3.1.2
yamlordereddictloader==0.4.0
zipp==3.15.0
